rename: 패키지 구조 수정
도메인 별 dto 패키지를 생성하여 dto 모아놓도록 수정
