+++ src/main/java/kr/hhplus/be/server/domain/point/dto/PointChargeResult.java
@@ -1,4 +1,6 @@
-package kr.hhplus.be.server.domain.point;
+package kr.hhplus.be.server.domain.point.dto;
+
+import kr.hhplus.be.server.domain.point.Point;
 
 public record PointChargeResult(
         Long id,
