+++ src/main/java/kr/hhplus/be/server/domain/point/dto/PointUseResult.java
@@ -1,4 +1,6 @@
-package kr.hhplus.be.server.domain.point;
+package kr.hhplus.be.server.domain.point.dto;
+
+import kr.hhplus.be.server.domain.point.Point;
 
 public record PointUseResult(
         Long id,
