+++ src/main/java/kr/hhplus/be/server/domain/product/ProductRepository.java
@@ -1,5 +1,6 @@
 package kr.hhplus.be.server.domain.product;
 
+import kr.hhplus.be.server.domain.product.dto.ProductSearchDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
