+++ src/main/java/kr/hhplus/be/server/domain/product/dto/ProductSearchDto.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.product;
+package kr.hhplus.be.server.domain.product.dto;
 
 public record ProductSearchDto(
     String productName,
