+++ src/test/java/kr/hhplus/be/server/domain/point/PointServiceTest.java
@@ -1,5 +1,9 @@
 package kr.hhplus.be.server.domain.point;
 
+import kr.hhplus.be.server.domain.point.dto.PointChargeResult;
+import kr.hhplus.be.server.domain.point.dto.PointHistorySearchResult;
+import kr.hhplus.be.server.domain.point.dto.PointSearchResult;
+import kr.hhplus.be.server.domain.point.dto.PointUseResult;
 import kr.hhplus.be.server.domain.point.exception.PointNotFoundException;
 import kr.hhplus.be.server.domain.user.User;
 import org.junit.jupiter.api.DisplayName;
