+++ src/test/java/kr/hhplus/be/server/domain/product/ProductServiceTest.java
@@ -1,5 +1,7 @@
 package kr.hhplus.be.server.domain.product;
 
+import kr.hhplus.be.server.domain.product.dto.ProductSearchDto;
+import kr.hhplus.be.server.domain.product.dto.ProductSearchResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
